feat: update development environment with new Dockerfile, install scripts, and workspace settings

+++ .devcontainer/post-create.sh
@@ -1,30 +1,44 @@
 #!/bin/sh
-cd /usr/local/redmine
 
-ln -s /workspaces/${PLUGIN_NAME} plugins/${PLUGIN_NAME}
-if [ -f plugins/${PLUGIN_NAME}/Gemfile_for_test ]
-then
-    cp plugins/${PLUGIN_NAME}/Gemfile_for_test plugins/${PLUGIN_NAME}/Gemfile 
+cd `dirname $0`
+cd ..
+BASEDIR=`pwd`
+PLUGIN_NAME=`basename $BASEDIR`
+
+if [ ! -f ~/.bashrc ]; then
+    cd ~/
+    tar xfz /.home.tgz
+    cd $BASEDIR
 fi
-cp plugins/${PLUGIN_NAME}/test/fixtures/*.yml test/fixtures
-ln -s /workspaces/${PLUGIN_NAME}/.devcontainer/launch.json .vscode/launch.json
 
-bundle install 
-bundle exec rake redmine:plugins:migrate
-bundle exec rake redmine:plugins:migrate RAILS_ENV=test
+if [ ! -f Gemfile ]; then
+    cp Gemfile_for_test Gemfile
+fi
+
+cd $REDMINE_ROOT
+
+git pull
+
+bundle install
+bundle exec rake redmine:plugins:ai_helper:setup_scm
 
 initdb() {
-    bundle exec rake db:create
+    if [ $DB != "sqlite3" ]
+    then
+        bundle exec rake db:create
+    fi
     bundle exec rake db:migrate
     bundle exec rake redmine:plugins:migrate
 
-    bundle exec rake db:drop RAILS_ENV=test
-    bundle exec rake db:create RAILS_ENV=test
+    if [ $DB != "sqlite3" ]
+    then
+        bundle exec rake db:drop RAILS_ENV=test
+        bundle exec rake db:create RAILS_ENV=test
+    fi
+
     bundle exec rake db:migrate RAILS_ENV=test
     bundle exec rake redmine:plugins:migrate RAILS_ENV=test
 }
-
-initdb
 
 export DB=mysql2
 export DB_NAME=redmine
@@ -42,4 +56,8 @@ export DB_PASSWORD=postgres
 export DB_HOST=postgres
 export DB_PORT=5432
 
+initdb
+
+rm -f db/redmine.sqlite3_test
+export DB=sqlite3
 initdb
